1、修改枚举值 2、删除日志线程

+++ TcpCommucation/ConnDisconnectEvt.cpp
@@ -8,7 +8,7 @@ namespace tc
 	ConnDisconnectEvt::ConnDisconnectEvt(TcpService* pSrv, SOCKET sendrecv/* = NULL*/)
 		:TcpEvt(pSrv, sendrecv)
 	{
-		evt = ETcpEvt::ConnDisconnect;
+		evt = static_cast<int>(ETcpEvt::ConnDisconnect);
 	}
 
 	ConnDisconnectEvt::~ConnDisconnectEvt()
